app/requirements.txt
tuya-connector-python
fastapi
uvicorn